[OPIK-6754] [FE] fix: move sortEnvironments to src/utils to satisfy layering
The api → shared edge tripped the `no-api-importing-components`
dep-cruiser rule once useEnvironmentsList pulled sortEnvironments from
shared/EnvironmentLabel/helpers.ts. Move the pure-data helper to
src/utils/environments.ts so the API layer can import it cleanly.
