+++ apps/opik-frontend/src/shared/EnvironmentLabel/helpers.ts
@@ -33,17 +33,3 @@ export const getDefaultEnvironmentMeta = (
 
 export const isDefaultEnvironmentName = (name: string | undefined | null) =>
   normalize(name) in DEFAULT_ENVIRONMENTS;
-
-const DEFAULT_ENV_SORT_PRIORITY: Record<string, number> = {
-  production: 0,
-  staging: 1,
-  development: 2,
-};
-
-export const sortEnvironments = <T extends { name: string }>(envs: T[]): T[] =>
-  envs.slice().sort((a, b) => {
-    const aPriority = DEFAULT_ENV_SORT_PRIORITY[normalize(a.name)] ?? Infinity;
-    const bPriority = DEFAULT_ENV_SORT_PRIORITY[normalize(b.name)] ?? Infinity;
-    if (aPriority !== bPriority) return aPriority - bPriority;
-    return a.name.localeCompare(b.name);
-  });
